[Comgr][hotswap] Give the operand resolver its own file
The operand resolver is a per-instruction accessor, not per-kernel state, and
every instruction format family takes one as its handler parameter. Sitting in
the raise context header it read like part of the context; it gets its own
header and source file, and its test moves along with it.
