+++ amd/comgr/src/hotswap/raiser/CMakeLists.txt
@@ -36,6 +36,7 @@ add_library(hotswap-raiser OBJECT
   user-sgpr-layout.cpp
   register-state.cpp
   raise-context.cpp
+  op-resolver.cpp
 )
 
 if(NOT TARGET hotswap::raiser)
